+++ src/constants.ts
@@ -30,3 +30,5 @@ export const GLOB_JSON5 = '**/*.json5';
 export const GLOB_YAML = '**/*.y?(a)ml';
 
 export const GLOB_TOML = '**/*.toml';
+
+export const GLOB_PACKAGE_JSON = '**/package.json';