+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- New config: `packageJson` via [`eslint-plugin-package-json`](https://www.npmjs.com/package/eslint-plugin-package-json).
+
 ## 0.1.4
 
 - New config: `json` (for linting .json, .jsonc, .json5 files) via [`eslint-plugin-jsonc`](https://www.npmjs.com/package/eslint-plugin-jsonc).